Refactor: type 이동
