+++ features/notification-list/container/NotificationListCont.tsx
@@ -2,21 +2,7 @@
 
 import { useEffect, useState } from 'react';
 import NotificationListPres from '../presentational/NotificationListPres';
-import { stripMarkdown } from '@/shared/utils/stripmarkdown';
-
-export type Notification = {
-  id: number;
-  type: number;
-  postsId: number;
-  senderId: string;
-  checkStatus: number;
-  createAt: string;
-  title: string;
-  content: string;
-  senderName: string;
-  senderProfileImg: string | null;
-  comment_content: string | null;
-};
+import { type Notification} from '../types'
 
 export default function NotificationListCont() {
   // 상위 props 대기
@@ -28,6 +14,7 @@ export default function NotificationListCont() {
   };
 
   const [notificationList, setNotificationList] = useState<Notification[]>([]);
+  
   useEffect(() => {
     fetch('/api/member/notifications')
       .then((res) => res.json())
