+++ features/notification-list/types.ts
@@ -1,11 +1,13 @@
-export type NotificationItemType = {
+export type Notification = {
   id: number;
-  userProfileImage?: string;
-  userNickname: string;
-  type: 'comment' | 'post';
+  type: number;
+  postsId: number;
+  senderId: string;
+  checkStatus: number;
+  createAt: string;
+  title: string;
   content: string;
-  postTitle: string;
-  postId: string;
-  date: string;
-  isRead: boolean;
+  senderName: string;
+  senderProfileImg: string | null;
+  comment_content: string | null;
 };
